+++ src/hocs/SceneBundle/actions.js
@@ -21,7 +21,7 @@ export function arenaLoadScene(
     isInitial,
     notifyAction: {
       arenaCurtainReducerKey: parentArenaReducerDict._curCurtain.reducerKey,
-      arenaSwitchVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
+      arenaCurtainVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
       sceneBundle,
       notifyData,
       isInitial
@@ -41,8 +41,10 @@ export function arenaLoadAsyncScene(
     asyncSceneBundle,
     isInitial,
     notifyAction: {
+      arenaSwitchReducerKey: parentArenaReducerDict._curSwitch.reducerKey,
+      arenaSwitchVReducerKey: parentArenaReducerDict._curSwitch.vReducerKey,
       arenaCurtainReducerKey: parentArenaReducerDict._curCurtain.reducerKey,
-      arenaSwitchVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
+      arenaCurtainVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
       asyncSceneBundle,
       notifyData,
       isInitial
@@ -59,8 +61,10 @@ export function sceneLoadStart(
 ) {
   return {
     type: ARENA_SCENEBUNDLE_LOAD_START,
+    arenaSwitchReducerKey: parentArenaReducerDict._curSwitch.reducerKey,
+    arenaSwitchVReducerKey: parentArenaReducerDict._curSwitch.vReducerKey,
     arenaCurtainReducerKey: parentArenaReducerDict._curCurtain.reducerKey,
-    arenaSwitchVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
+    arenaCurtainVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
     sceneBundle,
     asyncSceneBundle,
     notifyData,
@@ -77,7 +81,7 @@ export function sceneStartPlay(
   return {
     type: ARENA_SCENEBUNDLE_PLAY_START,
     arenaCurtainReducerKey: parentArenaReducerDict._curCurtain.reducerKey,
-    arenaSwitchVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
+    arenaCurtainVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
     sceneBundle,
     asyncSceneBundle,
     notifyData
@@ -93,7 +97,7 @@ export function sceneStopPlay(
   return {
     type: ARENA_SCENEBUNDLE_UNMOUNT_START,
     arenaCurtainReducerKey: parentArenaReducerDict._curCurtain.reducerKey,
-    arenaSwitchVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
+    arenaCurtainVReducerKey: parentArenaReducerDict._curCurtain.vReducerKey,
     sceneBundle,
     asyncSceneBundle,
     notifyData